common: Move GUI objects to the gui module
