+++ common.h
@@ -23,42 +23,6 @@ struct _FFTStuff {
 };
 extern FFTStuff fft;
 
-typedef struct _GuiObjs GuiObjs;
-struct _GuiObjs {
-  GtkWidget *button_abort;
-  GtkWidget *button_browse;
-  GtkWidget *button_clear;
-  GtkWidget *button_start;
-  GtkWidget *combo_card;
-  GtkWidget *combo_mode;
-  GtkWidget *entry_picdir;
-  GtkWidget *eventbox_img;
-  GtkWidget *frame_manual;
-  GtkWidget *frame_slant;
-  GtkWidget *grid_vu;
-  GtkWidget *iconview;
-  GtkWidget *image_devstatus;
-  GtkWidget *image_pwr;
-  GtkWidget *image_rx;
-  GtkWidget *image_snr;
-  GtkWidget *label_fskid;
-  GtkWidget *label_lastmode;
-  GtkWidget *label_utc;
-  GtkWidget *menuitem_about;
-  GtkWidget *menuitem_quit;
-  GtkWidget *spin_shift;
-  GtkWidget *statusbar;
-  GtkWidget *tog_adapt;
-  GtkWidget *tog_fsk;
-  GtkWidget *tog_rx;
-  GtkWidget *tog_save;
-  GtkWidget *tog_setedge;
-  GtkWidget *tog_slant;
-  GtkWidget *window_about;
-  GtkWidget *window_main;
-};
-extern GuiObjs   gui;
-
 extern GdkPixbuf *pixbuf_PWR;
 extern GdkPixbuf *pixbuf_SNR;
 extern GdkPixbuf *pixbuf_rx;
